main: log all dp urls that appear
uh, somehow, Google's developer preview links __quickly__ change
sometimes to one of the older previews (there's only one link on their
page, so it can't be the bots fault. This fixes that by logging __all__
url's that appear, so if it happens again, that will be the only time it
ever happens.

Signed-off-by: shagbag913 <sh4gbag913@gmail.com>